Revert constraints for add_column_with_default
There is a consistency issue with this: The outcome of the migration
(and whether or not we see a `NOT NULL` or a `CHECK` constraint) depends
on when the migration is being run and in which state the helper is at
that time. This includes a risk of seeing installations that have
different constraints, in particular differing from what we expect to
see in `db/structure.sql`.

As such, this commit reverts the line change.

A follow-up is
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/31736 which
deprecates the `add_column_with_default` helper anyways. For that, it's
helpful to keep using `NOT NULL` constraints (because the outcome is the
same if we just delegate to `add_column`).