Remove check for user being applicable codeowner
I mistakenly believed this to be desired behavior, and as part of
cleanup work ahead of a codeowner's project I "fixed" this. I had
thought that we wanted codeowner rules to NOT apply when the user was
listed as an owner of one of the applicable rules, essentially assuming
their approval as implied due to them taking the git action the check is
being applied to. However, this is incorrect, as we still want to push
users towards the MR workflow, even when they are in fact an owner who
_could_ approve the change.