Clear merge request error on push to source branch
If a merge error happens for some reason to a merge request, that error
will persist even after a new push to the repository. This makes it
impossible for a user to merge a merge request. The only workaround is
to close the existing merge request and open a new one.

If a user pushes to the source branch, we should clear out the merge
error since the diff would likely have changed. If a user pushes to the
target branch, we still want to keep the error present so the user can
take action. Clearing it out in that case would likely lead to confusion
because the error would disappear on unrelated changes.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/34186