Relax force pull mirror update restriction
We currently allow force pull mirror update via API and web UI if
the mirror wasn't updated within the last 5 minutes.

This affects pull mirrors that use SSH key authentication as they
will fail initially because the public key has to be added to the
source server. As a result, user have to wait 5 minutes before
they can retry. This also resulted to a QA spec to fail.

This change will fix it by not restricting force mirror updates
on mirrors that were never successfully updated.