Show correct last successful update timestamp
We were previously showing the `last_update_at` of pull mirror as
`Last successful update` timestamp. But that attribute gets set
whenever the mirror finished regardless if it failed or not.

We have the `last_successful_update_at` attribute that gets set
when the mirror successfully updated. This change uses that
attribute instead.