Add a specific tag to runners created during E2E tests
This is a simple work around to ensure that the runner will be found in
the first API page of results when trying to remove the runner after the
test ran. Ideally, the API retrieval should support pagination, but
settings tags is a best practice as well as more performant for
retrieving the runner via the API anyway.

Signed-off-by: Rémy Coutable <remy@rymai.me>