Skip section headers in legacy file parsing
If using the existing, legacy codeowners file parsing with a sectional
codeowners file, skip any headers. There is potential for this if
someone has a sectional codeowners file and the feature is disabled
(either they haven't enabled it yet, or have turned it off for some
reason..) With out this, we could potentially create new ApprovalRules
that would be meaningless.