Log the context metadata for rails requests
This logs the context for Rails-controller and Grape-API reqeusts
similar to what we do for Sidekiq.

The information comes from Labkit::Context.

This also avoids logging information for unpersisted groups and
projects in controllers. This was needed since the `#new` actions of
controllers would assign those ivars but the information is not
relevant for logs.