Expose secrets configuration if job has any
This builds on top of
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/31831
and expose secrets configuration when present to the runners.

This new `ci_secrets_management` feature is available for Premium and
above tiers.

At the moment the feature is hidden behind feature flag named
`ci_secrets_management_vault`.

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/28321.