Add a note about backporting columns with default
PostgreSQL 11 is required starting from 12.10. Before PostgreSQL 11,
adding a column requires a full table write, so if the table has a
considerable number of records, this causes for the table to be
unavailable until the column is added (which can take a while).

This commits adds a note on 'Add Column with Default' to state
backporting a column with default needs to use add_column_with_default,
and if a large table is involved, it should not add a default

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/207820