Hard-code non-configurable sampler intervals
While these samplers had intervals in 1_settings.rb, those weren't
exposed anywhere else: in the `gitlab.yml.example` file, in
`gitlab.rb.template` in Omnibus, or in our Helm charts.

Based on that, it doesn't appear that we expected people to configure
these intervals, so we should make the hard-coding explicit.