Add default values to _ide_theme_overrides
Add default values based on the white theme's colors to the
variables in _ide_theme_overrides.scss.

.ide.theme-dark is currently added to the top of the
_ide_theme_overrides.scss file to scope the refactor work of
moving to CSS variables to the dark theme only. So that other
themes, especially the white theme remain unaffected.

This MR adds default values to all the var() instances, so
that that scoping may be removed.