Do not replace methods of BatchLoader proxy objects
When the `replace_methods` option of BatchLoader#batch method is not
set as `false`, the gem is defining all the instance methods of loaded
object for the proxy instance of BatchLoader which can actually be an
order of magnitude slower than just falling back to `method_missing`.

Here we are setting `replace_methods` as false to disable this odd
behaviour.