Show storage limit alert for projects and groups
This adds an alert for projects and groups when the namespace
storage limit reaches a certain threhsold.
Depending on the threshold and alert_level we show different
styles of the alert.

We use cookies for to keep the alerts hidden for a user per
namespace and alert level. This is the first iteration of the
alert system for namespace storage limit and will be improved
later since it does not align with our UX guidelines to keep
alerts closed for a user.

We need to use cookies because we have no other system in place
so far to keep track of user + namespace + alert level.