Memoise database load balancing status
This sampler runs every few seconds, and in development mode, it can
flood the console with license load messages. This value can only change
during the lifetime of the sampler if someone uploads a new license, or
a license expires, but those are both pretty rare (and would need
associated configuration changes anyway).