Swap to db-tracking published issues
To have idempotent publishing jobs for publishing a project's
issues to the status page, issues are pushed based on the
confidential state of an issue. This commit swaps over to the
publishing an issue only if a StatusPage::PublishedIncident
exists for the issue. This allows finer control over to the user
on whether an issue should be published to the status page.