Add experimental internal Kubernetes API
- Return agent_id in response if present
- Receive token via Authorization header

Use standard way for Authorization which is the header. We currently not
care about the
Authorization type but may enforce this in future

Adds /project_info endpoint. Only for public projects. We will sort out
authorization for internal and private projects later

Remove need to hack around InternalHelpers. InternalHelpers assumes an
actor (we don't have it) and params (don't have that either). So get the
required data directly instead.