Add resolvers for the user <-> merge-request edges
These resolvers re-use the functionality of the merge_requests_resolver
to find merge requests either authored by or assigned to a user. The
author is the implicit argument of the edge, and they support all the
filtering arguments of the merge_requests_resolver, as well as being
able to restrict the project, either by ID or full-path.