Ignore title and description from services
The integration's title and description are not consistent between all
integrations. Some of them integrations allow the user to specify the
title or description but these values can always be the same.

Here, these values are hardcoded as virtual methods for each service, so
we can remove the column in the future.