Persist and send replication enabled fields
Secondary nodes are the source of truth for whether replication is
enabled or disabled to those nodes, because usually there is some way
for replication to be disabled via configuration on the secondary
machines.

E.g. if a secondary is configured to store a data type in
object storage, then replication is disabled, unless the Allow
replication to object storage checkbox is checked. If that checkbox is
unchecked, then the primary cannot know whether replication is enabled
or disabled for that secondary. Similarly, container repository must be
configured on the secondary; the primary cannot know whether it is
enabled or not.