Dont display subscription banner if auto_renew nil
In some instances we were showing the subscription banner when
auto_renew was nil. This was more of a data consistency issue and caused
alarm to a lot of customers and added to the support queue.

We've decided to not show the subscription banner if auto_renew is nil.

part of: https://gitlab.com/gitlab-org/growth/product/-/issues/1567