Rewrite repo size reduction docs
Purging data from a Git repo can be useful for different reasons,
including making cloning faster, or reducing repo size on the GitLab
server. The first is easy, the second is hard.

Documentation has been rewritten to use `git filter-repo` which is the
leading repo rewriting tool for this application, and is recommended by
the Git project.

Documentation now also provides workflow for pruning all kinds of refs,
including refs/keep-around/* which was not documented. This meant
purging files was incomplete from disk use perspective on the GitLab
server.

References to BFG have been removed.