Remove Sidekiq jobstr from logs and Sentry
When an exception happens in Sidekiq, Sentry gets a hash of the job
details, and then the same hash as a JSON-encoded string. That could be
useful if we had issues decoding the JSON, but that is very rare.
Sending this has a couple of problems:

1. It's wasteful.
2. It makes it hard to process any sensitive information out of the job,
   as we'd have to do so in both the hash and the string.

By removing the string, we only have to remove sensitive information
from the hash.