Add Rails/SaveBang Rubocop rule for tests
When updating an ActiveRecord object, result of create, update, destroy,
save should be checked whether if it is true or not.
It is not important to handle the result of update,
however we need to make sure of successful update.

Since this is just an introduction commit, some files will be excluded.