Remove need to call gitaly in ProjectPipelineStatus
Calling commit (gitaly) is un-nessary just to obtain the sha, for the
cache key. In fact, the cache key does not need to have the sha as the
sha is part of the cache value. The new cache key should be unique
enough for each project.

This saves us 20 gitaly calls on the dashboard page (assuming we have
one full page of projects).