Option to disable PAT expiry for self-managed instances
Adds an option in Admin Settings page to
toggle automatic expiry of PAT.
When enabled, token expires on the expiry date defined.
When disabled, token can continue to be used even after
expiry date.
If personal access token expiry policy is set, invalid
tokens will not be revoked in case the Automatic expiry
setting is disabled.
In a future MR, an in-app notification will be shown
to the user to update the token.