Translate legend in ProjectSelector component
The "legend" text in the ProjectSelector component was previously only
defined in the GlInfiniteScroll component from GitLab UI. There are two
problems with that: strings defined in GitLab UI aren't translatable,
and also the string is generic and referred to "items" rather than
"projects".

This also updates the existing `items` slot to use the Vue 2.6+ sytax.

Addresses https://gitlab.com/gitlab-org/gitlab/-/issues/217985.