Adjust spacing between ActiveToggle and next section
The toggle is a <label> element which is an inline-block so
the parent div has a strange height of 29px even though
the content is only 24px. I am using display: block and
removing the line-height to make sure the parent div is
exactly 24px.

This makes the spacing between this and the next section
correct at 1rem or 16px.