Ignore moment.js dependency for pikaday
Pikaday has a dependency on moment.js but moment.js
was removed from our codebase and build outputs
an error in the browser's console. Potential fix is to wait when
upstream removes moment.js as a direct dependency
but Pikaday looks unmaintained (last commit in 2018).
Another fix is to add IgnorePlugin from webpack
to ignore moment.js when processing Pikaday.