Remove allow_concurrency setting
Gitlab::Runtime.multi_threaded? now requires access to Gitlab.config
which isn't available early on in these initializers.

We don't really need this setting anyway because this just wraps the app
in a mutex so that it can only be called by one thread at a time.

We're already running Puma by default so we're confident that our
application works without this mutex.