Render security dashboard data for project if it has vulnerabilities
Previously we were relying on existence of pipeline to render the
security dashboard but with the standalone vulnerabilities dashboard
we will allow users to create their own vulnerabilities which do not
belong to any pipeline. To support this feature, we remove the pipeline
information from the dashboard and show the dashboard if it has any
vulnerabilities associated with itself.