Fix design overlay test case without adminNote
Here we improve the test case for when we are
trying to ensure that a user without adminNote
permission is unable to move an existing note.

Previously, we were testing arbitrary details of the
component. Here, we simplify to test only
the outcome w.r.t the DOM (in this case,
the element's `style` tag).