Use specialized worker for project share
Specialized workers use much less resource to update project
authorizations but it could lead to synchronization issues because it
doesn't serialize simultaneous update events. The previous approach is
not bulletproof either, inconsistencies do accumulate in the project
authorization cache over time. For now, we also use the previous
approach as a safety measure but with some delay and lower priority.