Log exceptions in Commits::CreateService
If an exception is caught in Commits::CreateService, the return value is
handled by the caller, but there is currently no way for an admin to
know what happened because there are many possibilities (e.g. validation
error, pre-receive error, etc.).

For better observability, this change logs the exception but does not
report it to Sentry because that could create noise (e.g. repository
size limits exceeded is not a bug in the system).

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/220168