Fix 500 error in BlobController#delete
If a validation or other error occurs in `BlobController#delete`, it
previously tried to render the `show` view, but that would render a 500
because the view assumes certain instance variables to be present
(e.g. `@last_commit`). Fix this by redirecting back to the blob in
question.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/220168