Preserve order when applying scoped label titles
Order matters when applying scoped labels in a way it does not when
applying regular labels. When applying scoped labels by id this order
is maintained but when applying by title, such as via the API, it is
not.

This change stores the intended order by title when provided as an
array or comma-separated string of label titles and applies it to the
label order when updating an issuable.