Fix Gitlab::Geo::Replication::BaseTransfer class
The uploader attribute was introduced is only used
to upload the file to Object Storage when the option
to sync object storage is enabled.

The problem is that we're setting this value during
the object initialization. It raises an error for
orphaned records while building the uploader because
model is nil.