Listing expired yet active tokens
When PAT expiry is not enforced,
(ref: https://gitlab.com/gitlab-org/gitlab/-/issues/214723)
the tokens that have expired yet continue
to be usable are listed in the Active
Personal Access Tokens section.
These tokens have a help text to indicate
that they have expired, yet the expiry is
not enforced. And, are given an option to revoke
the token.