Clear hard failure when updating mirror via UI
There was a regression wherein updating a mirror that hard failed
wasn't updating even when the update is triggered via web UI.

It was because the `StartPullMirroringService` was reseting the
retry count before calling `ProjectImportState#force_import_jobi!`.
As a result, the method will return and do nothing because it
detects that the mirror is due to be updated and it's no longer
considered as a hard failure.

To fix it, we remove the need to reset the retry count before
`#force_import_job!` is called. The retry count will be reset in
`#force_import_job!` once it deemed that the mirror can be force
updated.

Now, we only reset the retry count when updating the next
execution timestamp in the said service.