Enable Package File replication by default
We are keeping the feature flag around for at least one release, out of
caution, since it enables so much new code. Anyway, the locations of the
conditionals may be reused to add a feature allowing admins to disable
replication per replicable model per secondary:
https://gitlab.com/gitlab-org/gitlab/-/issues/222715

I renamed the feature flag because there are other usages of it for the
frontend development work, which needs to remain disabled by default for
now.