Introduce namespace limits table
The two initial columns will be used to determine how much extra
storage got purchased and how long it is valid. It is therefore
part of the total namespace storage limit.

We are not adding those two columns to the namespace table
directly because we want to reduce the amount of columns there.