Add a Danger message when Sidekiq queues change
When Sidekiq queues are added, or existing queues are changed, this can
have some impact on our infrastructure as we may have a change in
workload that leads to us needing to make an infrastructure change.

The Scalability team can help with this, so point to them as a first
step.