Document code design of approval rules
In this development document, major parts of the application being
used to implement the approval rules feature is explained in a
high level manner.

The goal is to help contributors understand the code design easier
and to also help those familiar with the codebase to see if there
are parts to improve.

This doesn't document the entire feature yet as it's missing other
parts specified in the TODO section of the document.