Render storage usage graph with a limit
When a limit is set, users should see how much of their storage
usage gets used in the storage usage graph. For the case when the
limit is either 0 or the current usage exceeds the limit, we just
render the distribution.

We will handle the design of exceeded usage at a later stage.