Add ci_build_metadata.secrets to ignored jsonb columns
This column does have a schema, but it's present (and so validated) in
EE only, whereas the column can exist in both CE and EE. The result is
that the spec fails in gitlab-foss, but not gitlab. This fix gets us
green again while we try to work out what to do.