Add batch processing to the repository indexing
Instead of enqueuing a Sidekiq job for each repository to be indexed,
we are now leveraging a Redis ZSET that is filled with Project to be
treated by a cron worker.

Every run, the worker pulls _N_ Projects from the queue, process them
and re-enqueue any failures.

This process enables us to leverage the Elasticsearch Bulk API more
effectively