Only load project license if needed
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/24310 introduced a
unintended side effect: instead of loading the project license name only
when the explore tab were present, the license would always be loaded
even if it were not displayed.

To fix that, we now load the license only if it is needed and memoize
the value via `SafeRequestStore` in the helper.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/223740