Matrix results can be displayed in vector charts
When a matrix result is obtained from prometheus, the parsing
defines a "values" array with a time range of results.

However, some charts can only present an instant vector.

This change enables all matrix type results to be visible on charts
that support instant vectors. To do so it picks the last result
and presents it as "the" current vector.