Lazy render diff line comment button
This changes the diff line comment button to only render
after the user hovers over the line for the first time.
On initial load the button doesn't exist in the DOM to save some memory,
once the user hovers over a line the button then gets added
into the DOM and on subsequent hovers it shows/hides
like previously.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/223077