Add migration for product analytics table
This table is part of the "product analytics experiment" and as such
marked "experimental". The goal here is to explore the product analytics
as a MVP feature more. We are explicitly not spending time on relational
modeling here.

We expect significant changes to the database part of this once the
feature has been validated. Therefore, we expect to drop the table
when feature validation is complete. All data will be lost.